+++ Foxtaur/Foxtaur.LibRenderer/Models/Ray.cs
@@ -31,15 +31,13 @@ public class Ray
     /// <summary>
     /// Casts ray from screen to 3D space
     /// </summary>
-    public static Ray CreateByScreenRaycasting(ICamera camera, float screenX, float screenY, float viewportWidth,
-        float viewportHeight)
+    public static Ray CreateByScreenRaycasting(ICamera camera, double screenX, double screenY, double viewportWidth, double viewportHeight)
     {
-        var normalizedDeviceX = screenX / viewportWidth * 2.0f - 1.0f;
-        var normalizedDeviceY = screenY / viewportHeight * 2.0f - 1.0f;
+        var normalizedDeviceX = screenX / viewportWidth * 2.0 - 1.0;
+        var normalizedDeviceY = screenY / viewportHeight * 2.0 - 1.0;
 
         var nearPoint = camera.Position3D.AsVector3();
-        var farPoint =
-            camera.BackProjectionMatrix.TransformPerspectively(new Vector3(normalizedDeviceX, normalizedDeviceY, 1.0f));
+        var farPoint = camera.BackProjectionMatrix.TransformPerspectively(new Vector3((float)normalizedDeviceX, (float)normalizedDeviceY, 1.0f));
 
         return new Ray(nearPoint.AsPlanarPoint3D(), farPoint.AsPlanarPoint3D());
     }
@@ -55,21 +53,21 @@ public class Ray
         var k = p0 - sphere.Center.AsVector3();
 
         var b = Vector3.Dot(k, rayDirection);
-        var c = Vector3.Dot(k, k) - (float)Math.Pow(sphere.Radius, 2);
-        var d = (float)Math.Pow(b, 2) - c; // a = 1 because rayDirection is normalized
+        var c = Vector3.Dot(k, k) - Math.Pow(sphere.Radius, 2);
+        var d = Math.Pow(b, 2) - c; // a = 1 because rayDirection is normalized
 
         if (d < 0)
         {
             return new List<PlanarPoint3D>();
         }
 
-        var dSqRt = (float)Math.Sqrt(d);
+        var dSqRt = Math.Sqrt(d);
 
-        var t0 = -1.0f * b + dSqRt;
-        var t1 = -1.0f * b - dSqRt;
+        var t0 = -1.0 * b + dSqRt;
+        var t1 = -1.0 * b - dSqRt;
 
-        var i0 = p0 + rayDirection * t0;
-        var i1 = p0 + rayDirection * t1;
+        var i0 = p0 + rayDirection * (float)t0;
+        var i1 = p0 + rayDirection * (float)t1;
 
         return new List<PlanarPoint3D>()
         {
